% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_visdata.R
\name{fit.visdata}
\alias{fit.visdata}
\title{This functino processes visual data to determine detection probabilities and strip widths for each platform
Saves model output to files, and makes plots for evaluating model fit.
Returns: 
tDist <- Truncation distances fit_visdata}
\usage{
fit.visdata(visData, spIdxON2, PLC, keyList1, adjList1)
}
\description{
This functino processes visual data to determine detection probabilities and strip widths for each platform
Saves model output to files, and makes plots for evaluating model fit.
Returns: 
tDist <- Truncation distances fit_visdata
}
\examples{
fit.visdata()
}
