% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missingdata.R
\name{plot.missingdata}
\alias{plot.missingdata}
\title{Function generates set of histograms showing data gaps by year.  
RETURNS:
  percFilled<- A table of percentage data available by year for each covariate in covarList}
\usage{
\method{plot}{missingdata}(mySegments, covarList, fNamePrefix, varUnits)
}
\description{
Function generates set of histograms showing data gaps by year.  
RETURNS:
  percFilled<- A table of percentage data available by year for each covariate in covarList
}
\examples{
plot.missingdata() 
}
