% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot.timeseries}
\alias{plot.timeseries}
\title{Function to plot acoustic timeseries data and save to image.}
\usage{
plot.timeseries(siteList, outDir, AcOnlySegments)
}
\description{
Function to plot acoustic timeseries data and save to image.
}
\examples{
plot.timeseries() 
}
