% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_covars.R
\name{transform.covars}
\alias{transform.covars}
\title{Function applies transformations to data based on string.}
\usage{
\method{transform}{covars}(mySegments, covarList, transformList)
}
\description{
Function applies transformations to data based on string.
}
\examples{
transform.covars() 
}
