% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cleveland.R
\name{plot.cleveland}
\alias{plot.cleveland}
\title{Function plots covariates to look for skew}
\usage{
\method{plot}{cleveland}(mySegments, covarList, isTrans, fNamePrefix,
  varUnits)
}
\description{
Function plots covariates to look for skew
}
\examples{
plot.cleveland() 
}
