% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covarDensity.R
\name{plot.covarDensity}
\alias{plot.covarDensity}
\title{Function plots presence absence as a function of predictor variables.}
\usage{
\method{plot}{covarDensity}(mySegments, covarList, presAbs, fNamePrefix,
  varUnits)
}
\description{
Function plots presence absence as a function of predictor variables.
}
\examples{
plot.covarDensity() 
}
